Some more layout stuff
Added a central hall, made it compatible with the whole outpost notoriety system
(except the entry teleport stuff, that needs to be fixed). Added graphics for
signs and stuff, marines, turrets, alarm lights, and a fancy particle effect for
the Level Transporter.

Still to do: Tons. Need to remember to export the rest of the textures used into
their own PNGs, for example. I'm typing this specifically so I don't forget it.